refactor: api 호출함수 분리
