+++ src/api/wrongAnswersNote/fetchUserWrongAnswers.ts
@@ -1,0 +1,20 @@
+import { createClient } from "@/utils/supabase/client";
+import { Tables } from "../../../database.types";
+
+type UserAnswerType = Tables<"user_answer">;
+
+const supabase = createClient();
+
+// 'user_answer' 테이블에서 틀린문제만 가져오는 함수 정의
+export const fetchUserWrongAnswers = async (userId: string): Promise<UserAnswerType[]> => {
+  const { data, error } = await supabase
+    .from("user_answer")
+    .select("*")
+    .eq("user_id", userId)
+    .eq("is_corrected", false);
+
+  if (error) {
+    throw new Error(error.message);
+  }
+  return data;
+};
